// MEAN stack project/frontend/src/app/models/registerRequest.ts
export class RegisterRequest{
    _id: string;
    ime: string;
    prezime: string;
    username: string;
    password: string;
    telefon: string;
    email: string;
    nazivPreduzeca: string;
    adresaPreduzeca: string;
    pib: string;
    maticniBroj: string;
    logo: string;
    status: string;
}